# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SponsorshipOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts for user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled, paused]
      responses:
        "200":
          description: Contracts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractWithExtraction"
  /contracts/upload:
    post:
      operationId: uploadContract
      tags: [contracts]
      summary: Upload contract PDF and trigger AI extraction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
              required: [file]
      responses:
        "201":
          description: Contract created, extraction pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract with extraction data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contract with extraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractWithExtraction"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /contracts/{id}/reprocess:
    post:
      operationId: reprocessContract
      tags: [contracts]
      summary: Re-trigger AI extraction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reprocessing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/extraction:
    patch:
      operationId: updateExtraction
      tags: [contracts]
      summary: Update AI extraction data (user corrections)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExtractionBody"
      responses:
        "200":
          description: Updated extraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractExtraction"

  # MILESTONES
...
components:
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled, paused]
        originalFilename:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
        aiProcessingStatus:
          type: string
          enum: [pending, processing, complete, failed]
        aiError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, title, status, aiProcessingStatus, createdAt, updatedAt]
    ContractExtraction:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        brandName:
          type: string
        creatorName:
          type: string
        campaignName:
          type: string
        totalPayment:
          type: string
        currency:
          type: string
        paymentSchedule:
          type: array
          items:
            type: object
        deliverables:
          type: array
          items:
            type: object
        deadlines:
          type: array
          items:
            type: object
        platforms:
          type: array
          items:
            type: string
        requiredHashtags:
          type: array
          items:
            type: string
        requiredMentions:
          type: array
          items:
            type: string
        requiredLinks:
          type: array
          items:
            type: object
        approvalRequirements:
          type: string
        revisionTerms:
          type: string
        usageRights:
          type: string
        exclusivityClauses:
          type: string
        latePaymentTerms:
          type: string
        cancellationTerms:
          type: string
        riskFlags:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        aiSummary:
          type: string
        editedByUser:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, contractId, createdAt, updatedAt]
    RiskFlag:
      type: object
      properties:
        flag:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        clause:
          type: string
      required: [flag, severity]
    ContractWithExtraction:
      type: object
      properties:
        contract:
          $ref: "#/components/schemas/Contract"
        extraction:
          $ref: "#/components/schemas/ContractExtraction"
        milestonesCount:
          type: integer
        notionExports:
          type: array
          items:
            type: object
      required: [contract]
    UpdateContractBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled, paused]
    UpdateExtractionBody:
      type: object
      properties:
        brandName:
          type: string
        creatorName:
          type: string
        campaignName:
          type: string
        totalPayment:
          type: string
        currency:
          type: string
        paymentSchedule:
          type: array
          items:
            type: object
        deliverables:
          type: array
          items:
            type: object
        requiredHashtags:
          type: array
          items:
            type: string
        requiredMentions:
          type: array
          items:
            type: string
        requiredLinks:
          type: array
          items:
            type: object
        riskFlags:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        aiSummary:
          type: string
...